requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.8
pytest==7.3.1
Pillow==9.5.0
sentinelsat==1.2.1
python-multipart==0.0.6
requests==2.31.0
pytest-dependency==0.5.1